Move images to 'images' directory
