+++ saleor/static/images/paypal-logo.png
